+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Added Spanish translation [macagua]
 
+- Use `pytest-plone` for testing.
+  [ericof]
+
 
 ## 1.0.0a1 (2022-12-06)